feat: add authentication protection to agent creation
- Hide 'New Agent' button for logged-out users on /agents page
- Protect /agents/new page with server-side authentication check
- Redirect unauthenticated users to sign-in page
