+++ apps/web/src/app/agents/page.tsx
@@ -4,10 +4,13 @@ import AgentsList from "@/components/AgentsList";
 import { prisma } from "@/infrastructure/database/prisma";
 import { PageLayout } from "@/components/PageLayout";
 import { sortAgentsByBadgeStatus } from "@/shared/utils/agentSorting";
+import { auth } from "@/infrastructure/auth/auth";
 
 export const dynamic = 'force-dynamic';
 
 export default async function AgentsPage() {
+  const session = await auth();
+  
   const dbAgents = await prisma.agent.findMany({
     where: {
       ephemeralBatchId: null, // Exclude ephemeral agents
@@ -49,14 +52,16 @@ export default async function AgentsPage() {
   return (
     <PageLayout>
       <div className="space-y-8">
-        <div className="flex justify-end">
-          <Link
-            href="/agents/new"
-            className="inline-flex items-center rounded-md border border-transparent bg-blue-600 px-4 py-2 text-sm font-medium text-white shadow-sm hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2"
-          >
-            New Agent
-          </Link>
-        </div>
+        {session?.user?.id && (
+          <div className="flex justify-end">
+            <Link
+              href="/agents/new"
+              className="inline-flex items-center rounded-md border border-transparent bg-blue-600 px-4 py-2 text-sm font-medium text-white shadow-sm hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2"
+            >
+              New Agent
+            </Link>
+          </div>
+        )}
         <AgentsList agents={sortedAgents} />
       </div>
     </PageLayout>
